+++ index.ts
@@ -2,7 +2,7 @@
  * Automatisch erzeugte globale index.ts Datei fuer Speech-Framework
  *
  * Konfiguration: stable
- * Version: 0.5.15.0059 (Beta) vom 19.10.2019
+ * Version: 0.5.16.0060 (Beta) vom 17.12.2019
  */