+++ src/cloud/google/google-config-data.interface.ts
@@ -1,10 +1,10 @@
 /**
  * GoogleConfigData-Schnittstelle
  *
- * API-Version: 1.0
- * Datum:       07.05.2019
+ * API-Version: 1.1
+ * Datum:       03.11.2019
  *
- * Letzte Aenderung: 07.05.2019
+ * Letzte Aenderung: 03.11.2019
  * Status: rot
  *
  * @module cloud/google
@@ -19,6 +19,8 @@
 export interface GoogleConfigDataInterface {
     /** legt den App Key fuer die Verbindung zum Server fest */
     googleAppKey: string;
+    /** legt die URL fuer den Google Tokenserver fest */
+    googleServerUrl: string;
     /** legt die URL fuer den Dialogflow Tokenserver fest */
     dialogflowTokenServerUrl: string;
     /** legt die   ProjektID fuer Dialogflow fest */